// template_23/store/services/path-request.enum.ts
export enum PathRequest {
  Registration = 'auth/registration',
  Login = 'auth/login',
  Update = 'auth',
  ResetPassword = 'auth/reset-password',
  GetVerifyCode = 'auth/get-verify-code',
  GetEmailVerifyCode = 'auth/email-verify-code',
  GetVerifyCodeMobileNumber = 'auth/code-profile-mobile',
  GetUserProfile = 'auth/user-with-profiles',
  CreateProfile = 'profiles',
  BlackList = 'profiles/blacklist',
  Order = 'profiles/order',
  AddProfileReview = 'profiles/review/profile',
  AddChildrenReview = 'profiles/review/children',
  GetReviewByAuthor = 'profiles/review/author',
  GetReviewByRecipient = 'profiles/review/recipient',
  GetReviewForProfileCard = 'profiles/review/card',
  GetProfileCard = 'profiles/card',
  SearchHousekeeper = 'profiles/search',
  ChooseProfile = 'profiles/current',
}